+++ src/main/java/com/example/devcoiff/Services/IService.java
@@ -6,7 +6,7 @@ import com.example.devcoiff.Entities.Service;
 import java.util.List;
 
 public interface IService {
-    public Service ajoutService(Service service, Integer idCoiff);
+    public Service ajoutService(Service service, Long idCoiff);
     public void modifierService(Service service);
     public void supprimerService(Integer id);
     public Service getService(Integer id);
